requirements.txt
boto3
pandas
requests
snowflake-connector-python
